fix(ru-client, ru-client-model): create controllers, initialize config yaml file, change some models

+++ fs-platform-ru/backend/domain/ru-client/src/main/kotlin/com/fs/client/repository/ClientRepository.kt
@@ -5,11 +5,11 @@ import reactor.core.publisher.Mono
 
 interface ClientRepository {
 
-    fun getClintById(id: Long): Mono<ClientModel>
+    fun getClintById(id: Int): Mono<ClientModel>
 
-    fun updateClientInfo(id: Long): Mono<Long>
+    fun updateClientInfo(id: Int): Mono<Long>
 
     fun createClient(clientModel: ClientModel): Mono<Long>
 
-    fun deleteClient(id: Long): Mono<Long>
+    fun deleteClient(id: Int): Mono<Long>
 }
